# bot_max_backend/admission-planning.openapi.yaml
openapi: 3.0.3
info:
  title: Admission planning — направления, подборка и план 5×5
  version: 0.2.0-draft
  description: |
    Целевой контракт. Все операции требуют JWT; sessionId = principal UUID,
    чужая/несуществующая сессия возвращает 404. Неизвестный ID — 404,
    некорректное тело/фильтр/несовместимые связи — 400, конфликт версий — 409.
    Direction — официальное направление; programId — конкретный вариант конкурса
    (вуз, кампус, направление, форма обучения, финансирование, тип конкурса, год).
    Все баллы и происхождение рекомендаций вычисляет сервер, клиент передаёт только ID.

    Ранжирование: доступное БВИ первым; затем eligible с известной разницей
    totalScore - passingScorePreviousYear по убыванию; затем eligible без исторических
    данных. При равенстве — totalScore DESC, universityId ASC, programId ASC.
    incomplete/ineligible без доступного БВИ видны в ручном поиске, но не участвуют в автозаполнении.
    Нельзя сравнивать разные шкалы/конкурсы; неизвестная разница = null, не ноль.
    БВИ и общий конкурс рассчитываются независимо; ИД ограничены правилами вуза.

    Генерация (спецификация «Автоплан поступления 5x5» от 30.09.2026). Кандидаты — все
    рассчитанные программы до пагинации. Допустима программа с доступным БВИ либо с
    допуском по ЕГЭ и scoreDifference >= -allowedDeficit (0, 10, 15, 20; по умолчанию 15);
    верхней границы нет. Программы направления ранжируются по относительному проходному
...

...
servers:
  - url: /v1
security:
  - bearerAuth: []
paths:
...
  /sessions/{sessionId}/application-plan:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    get:
      operationId: getApplicationPlan
      summary: Восстановить сохранённый список и пересчитать баллы без изменения состава
      responses:
        '200':
          description: Для нового пользователя version=0, universities=[], bviProgramId=null
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        default: {$ref: '#/components/responses/Error'}
    put:
      operationId: saveApplicationPlan
      summary: Атомарно сохранить любые изменения списка и приоритетов
      description: |
        Полная замена состава. Пустой universities очищает план. Удаление вуза удаляет
        все его позиции. Замена — новые ID на тех же местах. Перестановка массивов
        меняет приоритеты. Повторы, чужая программа, более 5 вузов/направлений — 400.
        Программа без допуска (нет ЕГЭ или балл ниже минимума) и без БВИ — 400.
        bviProgramId должен присутствовать в плане и давать БВИ;
        null снимает выбор. При изменении профиля потеря права показывается предупреждением.
        expectedVersion должен совпадать с текущей версией; иначе 409 без частичной записи.
        Успех увеличивает version на 1. Баллы всегда пересчитываются сервером.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - {$ref: '#/components/schemas/PlanComposition'}
                - type: object
                  required: [expectedVersion]
                  properties:
                    expectedVersion: {type: integer, minimum: 0}
      responses:
        '200':
          description: План сохранён
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Plan'}
        '409': {$ref: '#/components/responses/Error'}
        default: {$ref: '#/components/responses/Error'}
  /sessions/{sessionId}/application-plan/preview:
    parameters:
      - {$ref: '#/components/parameters/SessionId'}
    post:
      operationId: previewApplicationPlan
      summary: Сгенерировать новый план или дополнить текущий черновик, не сохраняя
      description: |
        generate строит план с нуля по выбранным направлениям (пустой выбор — 400).
        fill сохраняет все ID и их относительный порядок из basePlan, заполняет места
        внутри существующих вузов и затем добавляет вузы до лимита. Для fill basePlan
        обязателен, для generate запрещён. В fill также передать пустой basePlan для
        подбора только по ЕГЭ при пустом выборе направлений. Результат — черновик;
        для сохранения передать composition в PUT с известной версией сохранённого плана.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  additionalProperties: false
                  required: [mode]
                  properties:
                    mode: {type: string, enum: [generate]}
                    allowedDeficit: {$ref: '#/components/schemas/AllowedDeficit'}
                - type: object
                  additionalProperties: false
                  required: [mode, basePlan]
                  properties:
                    mode: {type: string, enum: [fill]}
                    basePlan: {$ref: '#/components/schemas/PlanComposition'}
                    allowedDeficit: {$ref: '#/components/schemas/AllowedDeficit'}
      responses:
        '200':
          description: Предложение и причины неполного заполнения
          content:
            application/json:
              schema: {$ref: '#/components/schemas/PlanPreview'}
        default: {$ref: '#/components/responses/Error'}
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    SessionId: {name: sessionId, in: path, required: true, schema: {type: string, format: uuid}}
...
  responses:
    Error:
      description: '400 validation_failed; 401 unauthorized; 404 not_found; 409 version_conflict'
      content:
        application/json:
          schema:
            type: object
            required: [code, message]
            properties:
              code: {type: string}
              message: {type: string}
...
  schemas:
    Direction:
      type: object
      required: [id, code, name]
      properties:
        id: {type: string}
        code: {type: string, example: '09.03.04'}
        name: {type: string, example: Программная инженерия}
...
    ProgramOption:
      type: object
      required: [programId, programName, universityId, universityName, direction, city, campus, campaignYear, funding, studyForm, competitionType, eligibility, bviAvailable, totalScore, passingScorePreviousYear, previousYear, scoreDifference, comparison, breakdown, reasons, dataSource]
      properties:
        programId: {type: string}
        programName: {type: string}
        universityId: {type: string}
        universityName: {type: string}
        direction: {$ref: '#/components/schemas/Direction'}
        city: {type: string}
        campus: {type: string, nullable: true}
        campaignYear: {type: integer}
        funding: {type: string, enum: [budget, paid]}
        studyForm: {type: string, enum: [full_time, part_time, correspondence]}
        competitionType: {type: string, enum: [general, special_quota, separate_quota, target_quota]}
        eligibility: {type: string, enum: [eligible, incomplete, ineligible], description: Допуск к указанному конкурсу; БВИ проверяется отдельно}
        bviAvailable: {type: boolean}
        totalScore: {type: integer, nullable: true, description: Балл указанного конкурса с ИД; null при неполных результатах}
        passingScorePreviousYear: {type: integer, nullable: true}
        previousYear: {type: integer, nullable: true}
        scoreDifference: {type: integer, nullable: true, description: totalScore минус проходной сопоставимого конкурса; null если сравнение невозможно}
        comparison: {type: string, enum: [bvi, above_previous, near_previous, below_previous, insufficient_data, ineligible]}
        dataSource: {type: string, enum: [demo, verified], description: Демо-правила явно обозначаются в UI}
        breakdown:
          type: array
          items:
            type: object
            required: [kind, name, originalScore, countedScore, applied, explanation]
            properties:
              kind: {type: string, enum: [exam, olympiad, achievement]}
              name: {type: string}
              originalScore: {type: integer, nullable: true}
              countedScore: {type: integer, nullable: true}
              applied: {type: boolean}
              explanation: {type: string}
        reasons: {type: array, items: {type: string}, description: 'Недостающие предметы, пороги, ограничения, причины отказа в льготах'}
    PlanComposition:
      type: object
      required: [universities, bviProgramId]
      properties:
        bviProgramId: {type: string, nullable: true, description: Одно выбранное место использования БВИ или null}
        universities:
          type: array
          maxItems: 5
          description: Уникальные universityId; порядок задаёт личный приоритет вузов
          items:
            type: object
            required: [universityId, programIds]
            properties:
              universityId: {type: string}
              programIds:
                type: array
                minItems: 1
                maxItems: 5
                uniqueItems: true
                description: Порядок приоритета; один programId на directionId внутри вуза. Все программы принадлежат этому вузу и одной кампании.
                items: {type: string}
    AllowedDeficit:
      type: integer
      enum: [0, 10, 15, 20]
      default: 15
      description: Насколько итоговый балл может быть ниже прошлогоднего проходного
    PlanUniversityRanking:
      type: object
      required: [universityId, universityName, hits, sumOfPlaces, firstPlaces, places, addedAsFill, retained, message]
      properties:
        universityId: {type: string}
        universityName: {type: string}
        hits: {type: integer, description: Попаданий в топ-3 по выбранным направлениям}
        sumOfPlaces: {type: integer}
        firstPlaces: {type: integer}
        places:
          type: array
          items:
            type: object
            required: [directionId, directionName, place]
            properties:
              directionId: {type: string}
              directionName: {type: string}
              place: {type: integer, minimum: 1, maximum: 3}
        addedAsFill: {type: boolean, description: Добавлен при доборе}
        retained: {type: boolean, description: Вуз из basePlan}
        message: {type: string}
    PlanPreview:
      type: object
      required: [composition, options, explanations, warnings, calculatedAt, nearPreviousThreshold]
      properties:
        composition: {$ref: '#/components/schemas/PlanComposition'}
        options: {type: array, items: {$ref: '#/components/schemas/ProgramOption'}, description: Расчёт каждого programId из composition}
        explanations:
          type: array
          items:
            type: object
            required: [programId, reason, message]
            properties:
              programId: {type: string}
              reason: {type: string, enum: [selected_direction, related_direction, auto_fill, retained]}
              message: {type: string}
        universityRanking:
          type: array
          description: Почему вузы в автоплане; в GET/PUT пустой
          items: {$ref: '#/components/schemas/PlanUniversityRanking'}
        warnings: {type: array, items: {type: string}, description: 'Недостаток кандидатов, непокрытые интересы, потеря права БВИ и другие ограничения'}
        calculatedAt: {type: string, format: date-time}
        nearPreviousThreshold: {type: integer, minimum: 0, description: '|scoreDifference| <= threshold означает near_previous; выше/ниже за пределами. Настройка backend.'}
    Plan:
      allOf:
        - {$ref: '#/components/schemas/PlanPreview'}
        - type: object
          required: [version, savedAt]
          properties:
            version: {type: integer, minimum: 0}
            savedAt: {type: string, format: date-time, nullable: true}